requirements.txt
streamlit
langchain
pymongo
python-dotenv
google-generativeai
faiss-cpu
crewai
crewai_tools